chapter-6-data-engineering/requirements.txt
# Core Dependencies
numpy>=1.24.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Data Ingestion
psycopg2-binary>=2.9.0
google-cloud-bigquery>=3.10.0
boto3>=1.28.0
requests>=2.31.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Text Processing
spacy>=3.5.0
nltk>=3.8.0

# NLP & Entity Extraction
# Run after install: python -m spacy download en_core_web_sm

# Embeddings
openai>=1.0.0
sentence-transformers>=2.2.0
cohere>=4.0.0
voyageai>=0.1.0

# Vector Stores
chromadb>=0.4.0
pinecone-client>=2.2.0
weaviate-client>=3.22.0

# Search & Retrieval
rank-bm25>=0.2.2

# Knowledge Graph
networkx>=3.1.0
neo4j>=5.9.0

# Pipeline Orchestration
apache-airflow>=2.7.0
celery>=5.3.0

# Monitoring & Quality
prometheus-client>=0.17.0

# Data Quality
great-expectations>=0.17.0  # Optional

# Utilities
tqdm>=4.65.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
ipython>=8.14.0
jupyter>=1.0.0

# Type Hints
types-requests>=2.31.0
types-PyYAML>=6.0.0

# Optional: Advanced Features
# faiss-cpu>=1.7.4  # For FAISS vector search
# torch>=2.0.0  # For PyTorch-based models
# transformers>=4.30.0  # For HuggingFace models
# scikit-learn>=1.3.0  # For ML utilities